Don't log config diffs until initialized - less clutter at startup
